+++ .gitignore
@@ -1,5 +1,5 @@
-# Logs
-logs
+# ───── Node.js General ─────
+node_modules/
 *.log
 npm-debug.log*
 yarn-debug.log*
@@ -7,20 +7,35 @@ yarn-error.log*
 pnpm-debug.log*
 lerna-debug.log*
 
-node_modules
-.serverless
-coverage
-dist
-dist-ssr
-*.local
+# ───── TypeScript / JavaScript ─────
+*.js
+!jest.config.js
+*.d.ts
 
-# Editor directories and files
-.vscode/*
-!.vscode/extensions.json
-.idea
-.DS_Store
+# ───── Build / Coverage ─────
+dist/
+dist-ssr/
+build/
+coverage/
+cdk.out/
+.cdk.staging/
+.serverless/
+
+# ───── Environment / Local / Temp ─────
+*.local
+.env
+.env.*
 *.suo
 *.ntvs*
 *.njsproj
 *.sln
 *.sw?
+
+# ───── IDEs and Editors ─────
+.vscode/
+!.vscode/extensions.json
+.idea/
+.DS_Store
+
+# ───── OS Artifacts ─────
+Thumbs.db
