+++ frontend/.serverless_plugins/serverless-single-page-app-plugin.js
