+++ frontend/public/mockServiceWorker.js
