+++ frontend/src/mocks/browser.ts
