+++ frontend/src/queries/products.ts
