+++ package.json
@@ -1,73 +1,10 @@
 {
-  "name": "my-store-app",
-  "version": "1.0.0",
+  "name": "your-project",
   "private": true,
+  "workspaces": ["frontend", "infra"],
   "scripts": {
-    "start": "vite",
-    "build": "tsc && vite build",
-    "preview": "npm run build && vite preview",
-    "test": "vitest",
-    "test:ui": "vitest --ui",
-    "test:coverage": "vitest run --coverage",
-    "lint": "eslint --ignore-path .gitignore --fix src",
-    "prettier": "prettier src --write",
-    "client:deploy": "sls client deploy --no-config-change --no-policy-change --no-cors-change",
-    "client:deploy:nc": "npm run client:deploy -- --no-confirm",
-    "client:build:deploy": "npm run build && npm run client:deploy",
-    "client:build:deploy:nc": "npm run build && npm run client:deploy:nc",
-    "cloudfront:setup": "sls deploy",
-    "cloudfront:domainInfo": "sls domainInfo",
-    "cloudfront:invalidateCache": "sls invalidateCloudFrontCache",
-    "cloudfront:build:deploy": "npm run client:build:deploy && npm run cloudfront:invalidateCache",
-    "cloudfront:build:deploy:nc": "npm run client:build:deploy:nc && npm run cloudfront:invalidateCache",
-    "cloudfront:update:build:deploy": "npm run cloudfront:setup && npm run cloudfront:build:deploy",
-    "cloudfront:update:build:deploy:nc": "npm run cloudfront:setup && npm run cloudfront:build:deploy:nc",
-    "serverless:remove": "sls remove"
-  },
-  "dependencies": {
-    "@emotion/react": "^11.9.3",
-    "@emotion/styled": "^11.9.3",
-    "@mui/icons-material": "^5.8.4",
-    "@mui/material": "^5.8.7",
-    "@mui/styles": "^5.8.7",
-    "@types/yup": "^0.29.14",
-    "axios": "^0.27.2",
-    "formik": "^2.2.9",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-query": "^3.39.1",
-    "react-router-dom": "^6.3.0",
-    "yup": "^0.32.11"
-  },
-  "devDependencies": {
-    "@testing-library/jest-dom": "^5.16.4",
-    "@testing-library/react": "^13.3.0",
-    "@testing-library/user-event": "^14.2.1",
-    "@types/react": "^18.0.15",
-    "@types/react-dom": "^18.0.6",
-    "@typescript-eslint/eslint-plugin": "^5.30.5",
-    "@typescript-eslint/parser": "^5.30.5",
-    "@vitejs/plugin-react": "^1.3.2",
-    "@vitest/ui": "^0.18.0",
-    "c8": "^7.11.3",
-    "eslint": "^8.19.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-plugin-prettier": "^4.2.1",
-    "eslint-plugin-react": "^7.30.1",
-    "jsdom": "^20.0.0",
-    "msw": "^0.43.1",
-    "prettier": "2.7.1",
-    "serverless": "^3.20.0",
-    "serverless-finch": "^4.0.0",
-    "serverless-s3-cleaner": "^2.0.1",
-    "typescript": "^4.7.4",
-    "vite": "^2.9.13",
-    "vitest": "^0.17.1"
-  },
-  "msw": {
-    "workerDirectory": "public"
-  },
-  "engines": {
-    "node": ">=14.0.0"
+    "build": "npm run build --workspaces",
+    "deploy": "cd infra && cdk deploy"
   }
 }
+
